Moving "allocate" to public namespace
